Refactor & cleanup (#5)
* add javadoc

* add javadoc and concerns

* add helper method, refactor hasMove

* add OthelloReporter to deal with reporting games

* add OthelloReporter reporting

* refactor, remove whitespace and TODOS

* add Player abstract class

* add OthelloController abstract class

* remove manual player character declaration, add javadoc

* add javadoc

* add conclusion for random vs greedy

* add javadoc

* refactor and add concerns

* remove TODOS

* refactor and remove unneeded checks

* add removed attribute from stater code

May be required for testing safe to re add

* add extra constructor from starter code

Adds previous constructor that was used in main methods in starter code. Re add just in case it is used to test game.

* change OthelloReporter to use static methods

Makes the methods static, and take an othello value, rather than having to store reporter for many classes.

* change play to simplify reporting

Changes play method so that it is consistent among all controller children. Allows for enabling or disabling reporting by overriding play.

* fix whitespace

* change local variable name

Change name to represent what it is better

* change ArrayList to List

Program to interfaces, not implementations.